오프닝 수정, 메뉴 수정, 상대 선택지 생성, 데이터 추가, 스프라이트 렌더러 구조 변경

+++ PuyoPuyoContents/PuyoText.h
@@ -32,7 +32,7 @@ public:
 protected:
 	virtual void BeginPlay() override;
 private:
-	const size_t MAXLEN = 18;
+	const int MAXLEN = 18;
 	std::vector<int> ChMap;
 	size_t TextLength;
 	std::vector<USpriteRendererComponent*> TextSpriteRenderers;
